# Only compile XDE/ubench with LTO enabled