--- a/.cargo/config.toml
+++ b/.cargo/config.toml
@@ -1,6 +1,6 @@
 [alias]
 xtask = "run --package xtask --"
-ubench = "bench --package opte-bench --bench userland --"
+ubench = "bench --package opte-bench --bench userland --profile release-lto --"
 kbench = "bench --package opte-bench --bench xde --"
 
 [env]