Before:
```bash
kyle@farme:~/gits/opte$ time cargo xtask install
...
   Compiling xtask v0.1.0 (/develop/gits/opte/xtask)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 15.26s
     Running `target/debug/xtask install`
Building opteadm (release).
...
   Compiling oxide-vpc v0.1.0 (/develop/gits/opte/lib/oxide-vpc)
   Compiling opte-ioctl v0.1.0 (/develop/gits/opte/lib/opte-ioctl)
    Finished `release` profile [optimized + debuginfo] target(s) in 54.47s
Building xde (release).
   Compiling xde v0.1.0 (/develop/gits/opte/xde)
    Finished `release` profile [optimized + debuginfo] target(s) in 56.70s
Linking xde kmod...
Building xde dev link helper (release).
   Compiling compiler_builtins v0.1.146
   Compiling core v0.0.0 (/home/kyle/.rustup/toolchains/nightly-2025-02-20-x86_64-unknown-illumos/lib/rustlib/src/rust/library/core)
   Compiling xde-link v0.1.0 (/develop/gits/opte/xde/xde-link)
   Compiling rustc-std-workspace-core v1.99.0 (/home/kyle/.rustup/toolchains/nightly-2025-02-20-x86_64-unknown-illumos/lib/rustlib/src/rust/library/rustc-std-workspace-core)
   Compiling alloc v0.0.0 (/home/kyle/.rustup/toolchains/nightly-2025-02-20-x86_64-unknown-illumos/lib/rustlib/src/rust/library/alloc)
    Finished `release` profile [optimized + debuginfo] target(s) in 19.75s
Command requires admin privileges to install xde kernel module. Continue? [yY] Exiting...

real    2m27.355s
user    7m27.327s
sys     0m45.371s
```

After:
```bash
kyle@farme:~/gits/opte$ time cargo xtask install
...
   Compiling xtask v0.1.0 (/develop/gits/opte/xtask)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 14.85s
     Running `target/debug/xtask install`
Building opteadm (release).
...
   Compiling oxide-vpc v0.1.0 (/develop/gits/opte/lib/oxide-vpc)
   Compiling opte-ioctl v0.1.0 (/develop/gits/opte/lib/opte-ioctl)
    Finished `release` profile [optimized + debuginfo] target(s) in 39.79s
Building xde (release).
...
   Compiling xde v0.1.0 (/develop/gits/opte/xde)
    Finished `release-lto` profile [optimized + debuginfo] target(s) in 56.61s
Linking xde kmod...
Building xde dev link helper (release).
   Compiling compiler_builtins v0.1.146
   Compiling core v0.0.0 (/home/kyle/.rustup/toolchains/nightly-2025-02-20-x86_64-unknown-illumos/lib/rustlib/src/rust/library/core)
   Compiling xde-link v0.1.0 (/develop/gits/opte/xde/xde-link)
   Compiling rustc-std-workspace-core v1.99.0 (/home/kyle/.rustup/toolchains/nightly-2025-02-20-x86_64-unknown-illumos/lib/rustlib/src/rust/library/rustc-std-workspace-core)
   Compiling alloc v0.0.0 (/home/kyle/.rustup/toolchains/nightly-2025-02-20-x86_64-unknown-illumos/lib/rustlib/src/rust/library/alloc)
    Finished `release` profile [optimized + debuginfo] target(s) in 19.91s
Command requires admin privileges to install xde kernel module. Continue? [yY]                                                                                                              Exiting...

real    2m12.931s
user    9m5.097s
sys     0m46.227s
```

Realistically, this drops the cost of an incremental compile from ~67s to 50s on my Helios workstation.

Closes #702.