@zeeshanlakhani 2025-03-20
Times are a bit closer for me on incremental, but def a bit faster with this PR.